requirements_rag.txt
# RAG System Dependencies

# Core
python-dotenv==1.0.0

# Database
neo4j==5.15.0
supabase==2.3.0
psycopg2-binary==2.9.9

# Embeddings & ML
google-generativeai==0.3.2
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
numpy==1.26.3

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Utilities
tqdm==4.66.1
asyncio==3.4.3
aiohttp==3.9.1

# Existing dependencies (from requirements.txt)
openpyxl==3.1.2
selenium==4.15.2
